Make it work using pandoc
Obviously, pandoc needs to be installed on your system.

There are no checks in place to validate the pandoc options or that
pandoc is installed.

Currently, the pandoc command is run syncrounously. This should be
changed.

Although pandoc will work with lots of different file types, only .md is

currently supported