Add the ability to have different file formats
The default template option should now be a file name without an
extention, e.g. "default", there is now another option names extention
that defaults to html.

In the content files, now don't specify an ectention. This will look for
the file with the extention in the specified template directory.

For my own site, this enabled me to have teh template set to "archive"
for an archive page. When producing html, this will look for
    archive.html, when producing json, it looks for archive.json